to resolve circular dependency by noficationService and fileUploadService

+++ src/main/java/io/github/fasset/fasset/kernel/storage/FileSystemStorageService.java
@@ -1,9 +1,11 @@
 package io.github.fasset.fasset.kernel.storage;
 
 import io.github.fasset.fasset.kernel.messaging.UploadNotificationService;
-import io.github.fasset.fasset.kernel.messaging.model.FileUploadNotification;
+import io.github.fasset.fasset.model.files.FileUpload;
 import io.github.fasset.fasset.kernel.util.StorageException;
 import io.github.fasset.fasset.kernel.util.StorageFileNotFoundException;
+import io.github.fasset.fasset.service.FileUploadService;
+import io.github.fasset.fasset.service.impl.FileUploadServiceImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -32,13 +34,13 @@ public class FileSystemStorageService implements StorageService {
 
     private final Path rootLocation;
 
-    @Autowired
-    @Qualifier("notificationService")
-    private UploadNotificationService notificationService;
+    private final FileUploadService fileUploadService;
+
 
     @Autowired
-    public FileSystemStorageService(StorageProperties storageProperties){
+    public FileSystemStorageService(StorageProperties storageProperties, @Qualifier("fileUploadService") FileUploadService fileUploadService){
         rootLocation = Paths.get(storageProperties.getLocation());
+        this.fileUploadService = fileUploadService;
     }
 
     /**
@@ -53,29 +55,43 @@ public class FileSystemStorageService implements StorageService {
         log.info("Storing file into the directory : {}",fileName);
 
         if(file.isEmpty()){
+
             throw new StorageException("Failed to store empty file : "+fileName);
-        }
-        if(fileName.contains("..")){
-            // This is a security check
-            throw new StorageException("Cannot store file with relative path outside current directory "+fileName);
-        }
 
-        try {
-            Files.copy(file.getInputStream(),this.rootLocation.resolve(fileName),
-                    StandardCopyOption.REPLACE_EXISTING);
-        } catch (IOException e) {
-            throw new StorageException("Failed to store file " + fileName, e);
-        }
+        } else {
 
-        notificationService.sendNotification(configureNotification(this.rootLocation.resolve(fileName).toString(),"Dec 2017"));
+            FileUpload fileUpload = configureFileUploadAttributes(this.rootLocation.resolve(fileName).toString(),"Dec 2017");
+
+            if(fileUploadService.theFileIsAlreadyUploaded(fileUpload)){
+
+                log.info("The file : {} has already been uploaded",fileUpload.getFileName());
+
+            } else {
+
+                if (fileName.contains("..")) {
+                    // This is a security check
+                    throw new StorageException("Cannot store file with relative path outside current directory " + fileName);
+                } else {
+
+                    try {
+                        Files.copy(file.getInputStream(), this.rootLocation.resolve(fileName),
+                                StandardCopyOption.REPLACE_EXISTING);
+                    } catch (IOException e) {
+                        throw new StorageException("Failed to store file " + fileName, e);
+                    }
+
+                    fileUploadService.recordFileUpload(fileUpload);
+                }
+            }
+        }
 
     }
 
-    private FileUploadNotification configureNotification(String fileName,String month) {
+    private FileUpload configureFileUploadAttributes(String fileName, String month) {
 
         log.info("Getting ready to notify server of the file uploaded : {}",fileName);
 
-        return new FileUploadNotification(fileName,month, LocalDateTime.now().toString());
+        return new FileUpload(fileName,month, LocalDateTime.now().toString());
     }
 
 
@@ -94,7 +110,7 @@ public class FileSystemStorageService implements StorageService {
         try {
             filePathStream = Files.walk(this.rootLocation,1)
                     .filter(path -> !path.equals(this.rootLocation))
-                    .map(path -> this.rootLocation.relativize(path));
+                    .map(this.rootLocation::relativize);
         } catch (IOException e) {
             throw new StorageException("Failed to read stored files", e);
         }
